+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package psen_scan_v2
 Forthcoming
 -----------
 * Enable separate building of cpp-lib; extract sources into subproject psen_scan_v2_standalone
+* Renames ScanRange and DefaultScanRange to improve usability
 * Contributors: Pilz GmbH and Co. KG
 
 0.1.4 (2021-01-13)